Handle the doc/make_doc case more robustly
This is a bit of a hack, unfortunately, but it seems
that some of the packages that use the doc/make_doc
script assume that the script is called from within
the directory GAPROOT/pkg/PKGNAME/doc.

We don't have GAP set up in GitHub Actions to work like
this, but we temporarily move the <doc> directory of the
package to the right place to fool the script.